move notification settings to a separate feature
